Fix code blow up issue.
`this orElse that` used to duplicate the code for `that`, one per nested
flatmap in `this` -- this caused an order-of-magnitude blowup when using
accept("Content-Length") since such an expression generated nested
flatmaps, one per character.

The fix works in general, though when it is not needed, it causes a
slight increase in code size. The fix uses the same pattern from
ParseResultCPSCond.

Thanks to the fix, the nested-flatmap-free accept alternative acceptStr
is no longer necessary.